agents/requirements.txt
openai-agents>=0.3
fastapi>=0.115,<1.0
uvicorn>=0.32,<1.0
python-dotenv>=1.0
pydantic>=2.10,<3
httpx>=0.27,<1
python-multipart>=0.0.6
